+++ Scroller/EditableWrapper.js
@@ -6,12 +6,13 @@ import Spotlight, {getDirection} from '@enact/spotlight';
 import {getContainersForNode} from '@enact/spotlight/src/container';
 import {getTargetByDirectionFromElement} from '@enact/spotlight/src/target';
 import Accelerator from '@enact/spotlight/Accelerator';
+import {getLastPointerPosition, setPointerMode} from '@enact/spotlight/src/pointer';
 import {Announce} from '@enact/ui/AnnounceDecorator';
 import Touchable from '@enact/ui/Touchable';
 import classNames from 'classnames';
 import IString from 'ilib/lib/IString';
 import PropTypes from 'prop-types';
-import {useCallback, useEffect, useRef} from 'react';
+import {useCallback, useEffect, useLayoutEffect, useRef} from 'react';
 
 import $L from '../internal/$L';
 
@@ -86,7 +87,7 @@ const EditableWrapper = (props) => {
 	const hideItemFuncRef = editable?.hideItemFuncRef;
 	const showItemFuncRef = editable?.showItemFuncRef;
 	const focusItemFuncRef = editable?.focusItemFuncRef;
-
+	const initialSelected = editable?.initialSelected;
 	const mergedCss = usePublicClassNames({componentCss, customCss, publicClassNames: true});
 
 	const dataSize = children?.length;
@@ -134,7 +135,7 @@ const EditableWrapper = (props) => {
 	});
 	const announceRef = useRef({});
 
-	mutableRef.current.hideIndex = editable?.hideIndex;
+	mutableRef.current.hideIndex = editable?.hideIndex || dataSize;
 
 	// Functions
 
@@ -195,7 +196,7 @@ const EditableWrapper = (props) => {
 	}, [dataSize]);
 
 	const startEditing = useCallback((item) => {
-		if (item.dataset.index) {
+		if (item?.dataset?.index && (!item.hasAttribute('disabled') || item.className.includes('hidden'))) {
 			item.classList.add(componentCss.selected, customCss.selected);
 			mutableRef.current.selectedItem = item;
 			mutableRef.current.focusedItem?.classList.remove(customCss.focused);
@@ -222,17 +223,18 @@ const EditableWrapper = (props) => {
 				return current;
 			}
 		}
+		return null;
 	}, [scrollContentRef]);
 
 	const focusItem = useCallback((target) => {
 		const itemNode = findItemNode(target);
-		if (focusItemFuncRef && itemNode && !mutableRef.current.selectedItem) {
+		if (itemNode && !mutableRef.current.selectedItem) {
 			mutableRef.current.focusedItem?.classList.remove(customCss.focused);
 			mutableRef.current.focusedItem = itemNode;
 			mutableRef.current.focusedItem?.classList.add(customCss.focused);
 			mutableRef.current.prevToIndex = Number(itemNode.style.order) - 1;
 		}
-	}, [customCss.focused, findItemNode, focusItemFuncRef]);
+	}, [customCss.focused, findItemNode]);
 
 	const handleClickCapture = useCallback((ev) => {
 		if (ev.target.className.includes('Button')) {
@@ -254,7 +256,9 @@ const EditableWrapper = (props) => {
 			// Finalize orders and forward `onComplete` event
 			const orders = finalizeOrders();
 			finalizeEditing(orders);
-			focusItem(ev.target);
+			if (selectItemBy === 'press') {
+				focusItem(ev.target);
+			}
 			mutableRef.current.needToPreventEvent = true;
 		} else {
 			const targetItemNode = findItemNode(ev.target);
@@ -264,10 +268,11 @@ const EditableWrapper = (props) => {
 					mutableRef.current.targetItemNode = targetItemNode;
 					startEditing(targetItemNode);
 				}
+				mutableRef.current.needToPreventEvent = true;
 			} else {
 				mutableRef.current.targetItemNode = targetItemNode;
+				mutableRef.current.needToPreventEvent = false;
 			}
-			mutableRef.current.needToPreventEvent = false;
 		}
 	}, [finalizeEditing, finalizeOrders, findItemNode, focusItem, selectItemBy, startEditing]);
 
@@ -504,12 +509,12 @@ const EditableWrapper = (props) => {
 	}, [scrollContainerHandle, scrollContentRef]);
 
 	const handleMouseMove = useCallback((ev) => {
-		if (mutableRef.current.selectedItem && Number(mutableRef.current.selectedItem.style.order) - 1 < mutableRef.current.hideIndex) {
-			const {clientX} = ev;
+		const {clientX} = ev;
+		mutableRef.current.lastMouseClientX = clientX;
 
+		if (mutableRef.current.selectedItem && Number(mutableRef.current.selectedItem.style.order) - 1 < mutableRef.current.hideIndex) {
 			const toIndex = getNextIndexFromPosition(clientX, 0.33);
 
-			mutableRef.current.lastMouseClientX = clientX;
 			mutableRef.current.lastInputType = 'mouse';
 			moveItems(toIndex);
 		}
@@ -545,7 +550,9 @@ const EditableWrapper = (props) => {
 				if (selectedItem) {
 					const orders = finalizeOrders();
 					finalizeEditing(orders);
-					focusItem(ev.target);
+					if (selectItemBy === 'press') {
+						focusItem(ev.target);
+					}
 					mutableRef.current.needToPreventEvent = true;
 
 					setTimeout(() => {
@@ -556,6 +563,7 @@ const EditableWrapper = (props) => {
 					}, completeAnnounceDelay);
 				} else if (selectItemBy === 'press') {
 					startEditing(targetItemNode);
+					mutableRef.current.needToPreventEvent = true;
 				}
 			} else if (repeat && targetItemNode && !mutableRef.current.timer && selectItemBy === 'longPress') {
 				mutableRef.current.timer = setTimeout(() => {
@@ -564,16 +572,24 @@ const EditableWrapper = (props) => {
 			}
 		} else if (is('left', keyCode) || is('right', keyCode)) {
 			if (selectedItem) {
-				if (mutableRef.current.lastKeyEventTargetElement?.getAttribute('role') !== 'button' && Number(selectedItem.style.order) - 1 < mutableRef.current.hideIndex) {
-					if (repeat) {
-						SpotlightAccelerator.processKey(ev, moveItemsByKeyDown);
-					} else {
-						SpotlightAccelerator.reset();
-						moveItemsByKeyDown(ev);
+				if (mutableRef.current.lastKeyEventTargetElement?.getAttribute('role') !== 'button') {
+					if (Number(selectedItem.style.order) - 1 < mutableRef.current.hideIndex) {
+						if (repeat) {
+							SpotlightAccelerator.processKey(ev, moveItemsByKeyDown);
+						} else {
+							SpotlightAccelerator.reset();
+							moveItemsByKeyDown(ev);
+						}
 					}
-
 					ev.preventDefault();
 					ev.stopPropagation();
+				}
+			} else {
+				const nextTarget = getTargetByDirectionFromElement(getDirection(keyCode), target);
+
+				// Check if focus leaves scroll container.
+				if (nextTarget && !getContainersForNode(nextTarget).includes(mutableRef.current.spotlightId) && !ev.repeat) {
+					reset();
 				}
 			}
 		} else if (is('up', keyCode) || is('down', keyCode)) {
@@ -581,7 +597,7 @@ const EditableWrapper = (props) => {
 				const nextTarget = getTargetByDirectionFromElement(getDirection(keyCode), target);
 
 				// Check if focus leaves scroll container.
-				if (!getContainersForNode(nextTarget).includes(mutableRef.current.spotlightId)) {
+				if (nextTarget && !getContainersForNode(nextTarget).includes(mutableRef.current.spotlightId)) {
 					Spotlight.move(getDirection(keyCode));
 
 					const orders = finalizeOrders();
@@ -591,8 +607,29 @@ const EditableWrapper = (props) => {
 					ev.stopPropagation();
 				}
 			}
+		} else if (is('cancel', keyCode)) {
+			if (!repeat) {
+				if (selectedItem) {
+					const orders = finalizeOrders();
+					finalizeEditing(orders);
+					if (selectItemBy === 'press') {
+						focusItem(ev.target);
+					}
+					mutableRef.current.needToPreventEvent = true;
+
+					setTimeout(() => {
+						announceRef?.current?.announce(
+							selectedItemLabel + $L('move complete'),
+							true
+						);
+					}, completeAnnounceDelay);
+
+					ev.preventDefault();
+					ev.stopPropagation();
+				}
+			}
 		}
-	}, [finalizeEditing, finalizeOrders, findItemNode, focusItem, moveItemsByKeyDown, selectItemBy, startEditing]);
+	}, [finalizeEditing, finalizeOrders, findItemNode, focusItem, moveItemsByKeyDown, reset, selectItemBy, startEditing]);
 
 	const handleKeyUpCapture = useCallback((ev) => {
 		mutableRef.current.lastKeyEventTargetElement = ev.target;
@@ -627,7 +664,7 @@ const EditableWrapper = (props) => {
 			mutableRef.current.centeredOffset = rtl ? bodyWidth - (item.getBoundingClientRect().right + container.scrollLeft) : item.getBoundingClientRect().left + container.scrollLeft;
 			wrapperRef.current?.style.setProperty('--item-width', mutableRef.current.itemWidth + 'px');
 		}
-	}, [scrollContainerHandle, scrollContentRef]);
+	}, [centered, dataSize, scrollContainerHandle, scrollContentRef]);
 
 	useEffect(() => {
 		mutableRef.current.spotlightId = scrollContainerRef.current && scrollContainerRef.current.dataset.spotlightId;
@@ -646,25 +683,25 @@ const EditableWrapper = (props) => {
 		};
 	}, [handleMouseLeave, scrollContainerRef]);
 
-	useEffect(() => {
+	useLayoutEffect(() => {
 		if (removeItemFuncRef) {
 			removeItemFuncRef.current = removeItem;
 		}
 	}, [removeItem, removeItemFuncRef]);
 
-	useEffect(() => {
+	useLayoutEffect(() => {
 		if (hideItemFuncRef) {
 			hideItemFuncRef.current = hideItem;
 		}
 	}, [hideItem, hideItemFuncRef]);
 
-	useEffect(() => {
+	useLayoutEffect(() => {
 		if (showItemFuncRef) {
 			showItemFuncRef.current = showItem;
 		}
 	}, [showItem, showItemFuncRef]);
 
-	useEffect(() => {
+	useLayoutEffect(() => {
 		if (focusItemFuncRef) {
 			focusItemFuncRef.current = focusItem;
 		}
@@ -689,13 +726,34 @@ const EditableWrapper = (props) => {
 			}
 		};
 
-		scrollContentNode.addEventListener('scroll', handleMoveItemsByScroll);
+		setTimeout(() => {
+			scrollContentNode.addEventListener('scroll', handleMoveItemsByScroll);
+		}, 400); // Wait for finishing scroll animation when initial selected item is given.
 
 		return () => {
 			scrollContentNode.removeEventListener('scroll', handleMoveItemsByScroll);
 		};
 
 	}, [getNextIndexFromPosition, moveItems, scrollContainerHandle, scrollContentRef]);
+
+	useEffect(() => {
+		if (initialSelected?.itemIndex) {
+			scrollContainerHandle.current?.scrollTo({animate:false, position: {x: initialSelected.scrollLeft}});
+		}
+	}, [initialSelected?.itemIndex, initialSelected?.scrollLeft, scrollContainerHandle]);
+
+	useLayoutEffect(() => {
+		if (initialSelected?.itemIndex) {
+			const initialSelectedItem = wrapperRef.current.children[initialSelected.itemIndex - 1];
+			if (initialSelectedItem?.dataset.index) {
+				mutableRef.current.focusedItem = initialSelectedItem;
+				mutableRef.current.lastMouseClientX = getLastPointerPosition().x;
+				startEditing(initialSelectedItem);
+				setPointerMode(false);
+				Spotlight.focus(initialSelectedItem.children[1]);
+			}
+		}
+	}, []); // eslint-disable-line react-hooks/exhaustive-deps
 
 	return (
 		<TouchableDiv