+++ samples/sampler/stories/qa/IconItem.js
@@ -8,6 +8,7 @@ import {action} from '@enact/storybook-utils/addons/actions';
 import {boolean, number, select} from '@enact/storybook-utils/addons/controls';
 import ri from '@enact/ui/resolution';
 import {ScrollerBasic as UiScrollerBasic} from '@enact/ui/Scroller';
+import Touchable from '@enact/ui/Touchable';
 import classNames from 'classnames';
 import {useCallback, useEffect, useLayoutEffect, useRef, useState} from 'react';
 
@@ -64,7 +65,8 @@ const populateItems = ({index}) => {
 			else if (index === 6) return 'Gallery';
 		})(),
 		labelColor: index === 6 ? 'dark' : null,
-		labelOn: index === 6 ? 'focus' : null
+		labelOn: index === 6 ? 'focus' : null,
+		disabled: index === 1
 	};
 
 	return {index, iconItemProps};
@@ -76,6 +78,7 @@ for (let i = 0; i < 20; i++) {
 
 const Container = SpotlightContainerDecorator('div');
 const ContainerDivWithLeaveForConfig = SpotlightContainerDecorator({leaveFor: {left: '', right: ''}}, 'div');
+const TouchableDiv = Touchable('div');
 
 export const EditableIcon = (args) => {
 	const dataSize = args['editableDataSize'];
@@ -87,7 +90,9 @@ export const EditableIcon = (args) => {
 	const focusItem = useRef();
 	const divRef = useRef();
 	const mutableRef = useRef({
-		hideIndex: null
+		hideIndex: null,
+		initialSelected: {},
+		timer: null
 	});
 
 	useLayoutEffect(() => {
@@ -99,8 +104,19 @@ export const EditableIcon = (args) => {
 		mutableRef.current.hideIndex = dataSize;
 	}, [dataSize]);
 
+	const findItemNode = useCallback((node) => {
+		for (let current = node; current !== divRef.current && current !== document; current = current.parentNode) {
+			if (current.dataset.index) {
+				return current;
+			}
+		}
+	}, [divRef]);
+
 	const onClickModeButton = useCallback(() => {
 		setEditMode(mode => !mode);
+		mutableRef.current.initialSelected.scrollLeft = 0;
+		mutableRef.current.initialSelected.itemIndex = null;
+		mutableRef.current.timer = null;
 	}, [setEditMode]);
 
 	const onClickRemoveButton = useCallback((ev) => {
@@ -130,6 +146,49 @@ export const EditableIcon = (args) => {
 		}
 	}, []);
 
+	const handleHoldStart = useCallback(() => {
+		setEditMode(true);
+	}, [setEditMode]);
+
+	const handleMouseDown = useCallback((ev) => {
+		if (!ev.target.className.includes('Button')) {
+			const targetItemNode = findItemNode(ev.target);
+			if (targetItemNode && targetItemNode.style.order) {
+				mutableRef.current.initialSelected.itemIndex = targetItemNode.style.order;
+			}
+		}
+	}, [findItemNode]);
+
+	const handleKeyDown = useCallback((ev) => {
+		const {keyCode, repeat, target} = ev;
+		if (is('enter', keyCode) && target.getAttribute('role') !== 'button') {
+			if (repeat && !mutableRef.current.timer) {
+				const targetItemNode = findItemNode(ev.target);
+				if (targetItemNode && targetItemNode.style.order) {
+					mutableRef.current.initialSelected.itemIndex = targetItemNode.style.order;
+				}
+				mutableRef.current.timer = setTimeout(() => {
+					setEditMode(true);
+				}, 200);
+
+			}
+		}
+	}, [findItemNode]);
+
+	const handleKeyUp = useCallback((ev) => {
+		if (ev.target.getAttribute('role') === 'button') {
+			return;
+		}
+		if (mutableRef.current.timer) {
+			clearTimeout(mutableRef.current.timer);
+			mutableRef.current.timer = null;
+		}
+	}, []);
+
+	const handleScroll = useCallback((ev) => {
+		mutableRef.current.initialSelected.scrollLeft = ev.scrollLeft;
+	}, []);
+
 	const handleComplete = useCallback((ev) => {
 		const {orders, hideIndex} = ev;
 		mutableRef.current.hideIndex = hideIndex;
@@ -142,20 +201,22 @@ export const EditableIcon = (args) => {
 		});
 
 		for (let i = 0; i < orders.length; i++) {
-			newItems[i].disabled = (i >= hideIndex);
+			newItems[i].hidden = (i >= hideIndex);
 		}
 
 		setItems(newItems);
 	}, [items]);
 
 	useEffect(() => {
-		divRef.current.addEventListener('keydown', (ev) => {
-			const {keyCode} = ev;
-			if (isCancel(keyCode)) {
+		document.addEventListener('keydown', (ev) => {
+			if (isCancel(ev.keyCode)) {
 				setEditMode(false);
+				mutableRef.current.initialSelected.scrollLeft = 0;
+				mutableRef.current.initialSelected.itemIndex = null;
+				mutableRef.current.timer = null;
 			}
 		});
-	}, [divRef]);
+	}, []);
 
 	return (
 		<div ref={divRef}>
@@ -173,6 +234,7 @@ export const EditableIcon = (args) => {
 							hideItemFuncRef: hideItem,
 							showItemFuncRef: showItem,
 							focusItemFuncRef: focusItem,
+							initialSelected: mutableRef.current.initialSelected,
 							selectItemBy: 'press'
 						}}
 						focusableScrollbar={args['focusableScrollbar']}
@@ -191,49 +253,57 @@ export const EditableIcon = (args) => {
 						{
 							items.map((item, index) => {
 								return (
-									<div key={item.index} className={classNames(css.itemWrapper, {[css.hidden]: item.disabled})} aria-label={`Image ${item.index}`} data-index={item.index} style={{order: index + 1}}>
+									<div key={item.index} className={classNames(css.itemWrapper, {[css.hidden]: item.hidden})} aria-label={`Image ${item.index}`} data-index={item.index} style={{order: index + 1}} disabled={item.iconItemProps['disabled'] || item.hidden}>
 										<ContainerDivWithLeaveForConfig className={css.removeButtonContainer}>
-											{item.disabled ? null : <Button aria-label="Delete" className={css.removeButton} onClick={onClickRemoveButton} icon="trash" />}
-											{item.disabled ? null : <Button aria-label="Hide" className={css.removeButton} onClick={onClickHideButton} icon="minus" />}
-											{item.disabled ? <Button aria-label="Show" className={css.removeButton} onClick={onClickShowButton} icon="plus" /> : null}
+											{item.hidden ? null : <Button aria-label="Delete" className={css.removeButton} onClick={onClickRemoveButton} icon="trash" />}
+											{item.hidden ? null : <Button aria-label="Hide" className={css.removeButton} onClick={onClickHideButton} icon="minus" />}
+											{item.hidden ? <Button aria-label="Show" className={css.removeButton} onClick={onClickShowButton} icon="plus" /> : null}
 										</ContainerDivWithLeaveForConfig>
 										<IconItem
+											{...item.iconItemProps}
 											aria-label={`Image ${item.index}. Edit mode to press and hold OK key`}
-											className={css.iconItem}
-											disabled={item.disabled}
+											className={css.editableIconItem}
+											disabled={item.iconItemProps['disabled'] || item.hidden}
 											onClick={action('onClickItem')}
 											onFocus={onFocusItem}
-											{...item.iconItemProps}
 										/>
 									</div>
 								);
 							})
 						}
 					</Scroller> :
-					<Scroller
-						direction="horizontal"
-						onClick={action('onClickScroller')}
-						onKeyDown={action('onKeyDown')}
-						onScrollStart={action('onScrollStart')}
-						onScrollStop={action('onScrollStop')}
+					<TouchableDiv
+						onHoldStart={handleHoldStart}
+						onMouseDown={handleMouseDown}
+						onKeyDown={handleKeyDown}
+						onKeyUp={handleKeyUp}
 					>
-						<div className={classNames(css.scrollerWrapper, css.wrapper, {[css.centered]: args['editableCentered']})}> {
-							items.map((item, index) => {
-								return (
-									<div key={item.index} className={classNames(css.itemWrapper, {[css.hidden]: item.disabled})} aria-label={`Image ${item.index}`} data-index={item.index} style={{order: index + 1}}>
-										<div className={css.removeButtonContainer} />
-										<IconItem
-											aria-label={`Image ${item.index}. Edit mode to press and hold OK key`}
-											className={css.iconItem}
-											onClick={action('onClickItem')}
-											disabled={item.disabled}
-											{...item.iconItemProps}
-										/>
-									</div>
-								);
-							})}
-						</div>
-					</Scroller>
+						<Scroller
+							direction="horizontal"
+							onClick={action('onClickScroller')}
+							onKeyDown={action('onKeyDown')}
+							onScroll={handleScroll}
+							onScrollStart={action('onScrollStart')}
+							onScrollStop={action('onScrollStop')}
+						>
+							<div className={classNames(css.scrollerWrapper, css.wrapper, {[css.centered]: args['editableCentered']})}> {
+								items.map((item, index) => {
+									return (
+										<div key={item.index} className={classNames(css.itemWrapper, {[css.hidden]: item.hidden})} aria-label={`Image ${item.index}`} data-index={item.index} style={{order: index + 1}}>
+											<div className={css.removeButtonContainer} />
+											<IconItem
+												{...item.iconItemProps}
+												aria-label={`Image ${item.index}. Edit mode to press and hold OK key`}
+												className={css.iconItem}
+												disabled={item.iconItemProps['disabled'] || item.hidden}
+												onClick={action('onClickItem')}
+											/>
+										</div>
+									);
+								})}
+							</div>
+						</Scroller>
+					</TouchableDiv>
 				}
 			</Container>
 		</div>
@@ -256,9 +326,6 @@ export const EditableIconWithLongPress = (args) => {
 	const dataSize = args['editableDataSize'];
 	const [items, setItems] = useState(itemsArr);
 	const removeItem = useRef();
-	const mutableRef = useRef({
-		hideIndex: null
-	});
 
 	useLayoutEffect(() => {
 		itemsArr = [];
@@ -266,7 +333,6 @@ export const EditableIconWithLongPress = (args) => {
 			itemsArr.push(populateItems({index: i}));
 		}
 		setItems(itemsArr);
-		mutableRef.current.hideIndex = dataSize;
 	}, [dataSize]);
 
 	const onClickRemoveButton = useCallback((ev) => {
@@ -295,7 +361,6 @@ export const EditableIconWithLongPress = (args) => {
 			editable={{
 				centered: args['editableCentered'],
 				css,
-				hideIndex: mutableRef.current.hideIndex,
 				onComplete: handleComplete,
 				removeItemFuncRef: removeItem
 			}}