+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [2.7.5] - 2023-08-04
+
+### Added
+
+- `sandstone/Scroller` prop `editable.initialSelected` to allow start edit mode with selected item
+
+### Changed
+
+- `sandstone/Scroller` back key behavior to match the latest UX when `editable` is given
+
+### Fixed
+
+- `sandstone/Scroller` to handle focus moving properly when `editable` is given
+- `sandstone/Scroller` to not select disabled item in scroller when `editable` is given
+- `sandstone/VirtualList` to move focus properly by 5-way directional key hold when `spotlight/SpotlightContainerDecorator` config option `continue5WayHold` is set
+
 ## [2.7.4] - 2023-07-19
 
 ### Fixed